+++ src/components/Layout1.jsx
@@ -2,6 +2,7 @@ import Header from './Header1';
 import '../styles/Layout1.css';
 
 //Layout Padrão com Título e Subtítulo
+//É possível dizer qual o subtitulo a partir do parâmetro
 function Layout1({ subtitulo }) {
   return (
     <>
