+++ src/components/TagBox.jsx
@@ -1,6 +1,7 @@
 import '../styles/TagBox.css';
-import Icone from './Icone.jsx'
+import Icone from './Icone';
 
+//Box que apresenta tags populares
 function TagBox({ icone, nomeTag, nPublicacoes, corFundo, corDestaque }) {
   return (
     <button
