+++ src/main.jsx
@@ -2,10 +2,10 @@ import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import './styles/reset.css';
 import './styles/index.css';
-import Layout3 from './components/Layout3.jsx';
+import Layout4 from './components/Layout4.jsx';
 
 createRoot(document.getElementById('root')).render(
   <BrowserRouter>
-    <Layout3 />
+    <Layout4 />
   </BrowserRouter>
 );
