+++ src/styles/Layout3.css
@@ -1,3 +1,3 @@
 #root {
-    width: 100%;
-  }
+  width: 100%;
+}
