+++ src/styles/Layout4.css
@@ -1,0 +1,7 @@
+#root {
+  width: 100%;
+}
+
+.conteudo {
+  display: flex;
+}
